requirements.txt
pandas>=2.2
supabase>=2.0
python-dotenv>=1.0
requests>=2.31
xero-python
pyarrow>=15.0
pendulum>=3.0
pytest>=8.0
psycopg[binary]>=3.1
streamlit>=1.37
plotly>=5.22
altair>=5.3
fpdf2>=2.7.0
fastapi>=0.109
uvicorn>=0.27
sendgrid
google-generativeai
